direct solver / block allocator

+++ ngstd/stringops.cpp
@@ -7,17 +7,17 @@ namespace ngstd
   bool StringFitsPattern(const string & str, const string & pattern)
   {
     int patternpos = 0;
-    size_t strpos = 0;
+    int strpos = 0;
 
-    size_t patternsize = -1;
-    for(unsigned int i=0; patternsize < 0 && i < pattern.size(); i++)
+    int patternsize = -1;
+    for(int i=0; patternsize < 0 && i<pattern.size(); i++)
       if(pattern[i] == 0)
 	patternsize = i;
     if(patternsize < 0)
       patternsize = pattern.size();
 
-    size_t strsize = -1;
-    for(unsigned int i=0; strsize < 0 && i<str.size(); i++)
+    int strsize = -1;
+    for(int i=0; strsize < 0 && i<str.size(); i++)
       if(str[i] == 0)
 	strsize = i;
     if(strsize < 0)
@@ -32,7 +32,7 @@ namespace ngstd
 	int nextpartsize = 0;
 	int mindistance = 0;
 	int maxdistance = 0;
-	size_t strdistance;
+	int strdistance;
 	
 	while(patternpos < patternsize &&
 	      (pattern[patternpos] == '?' ||
@@ -61,7 +61,7 @@ namespace ngstd
 	
 
 
-	size_t foundpos;
+	int foundpos;
 	
 	if(nextpartsize == 0)
 	  foundpos = strsize;
